If Only...
The ocean was evermore still this witching hour. Overhead the moon radiates across the deep blues. It was only at a sloths pace as that encompassing sunlight dipped further in the deep blues. I was glad this day the waves were still. Anymore crashes would force me out this slub. And, looking onwards at the microscopic sands helps me to even momentarily, and even emotionally for a time feel connected towards those tightly compact abysses. But today the ocean's waves laid still for my hopes and dreams to have a moment of rest.