Added Enemy AI
Essentially I designated it a certain amount of health, damage, speed, attack distance, fovAngle, and view distance. I made these adjustable in the game. I made the enemy script find the game object with the player tag and if the health is equal to the variable amount which is the number of damage then the enemy will die. The player will move when seeing you. I used ray casting to check if players within field of view and yea.